feat: added device_and_app_management_windows_update_catalog_item datasource with examples
