+++ templates/data-sources/graph_beta_device_and_app_management_windows_update_catalog_item.md.tmpl
@@ -1,0 +1,17 @@
+---
+page_title: "{{.Name}} {{.Type}} - {{.ProviderName}}"
+subcategory: "Intune"
+description: |-
+{{ .Description | plainmarkdown | trimspace | prefixlines "  " }}
+---
+
+# {{.Name}} ({{.Type}})
+
+The Microsoft 365 Intune windows update catalog item data source provides information about a windows updates. Can be filtered by
+id, display_name , end_of_support_date and release_date_time.
+
+## Example Usage
+
+{{ tffile "examples/microsoft365_graph_beta/microsoft365_graph_beta_device_and_app_management_windows_update_catalog_item/datasource.tf" }}
+
+{{ .SchemaMarkdown | trimspace }}
